Allow changing of session cookie expiry.
Add a set_session_cookie_expire function that lets you change the expiry
for the current session's cookie. You can set a number of seconds, 0 to
expire the cookie when the browser quits or undef to fallback to the
configured defaults.

Written by Edmund von der Burg, tweaked by me.